adding state backup for rateplan and refactoring
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C36486

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SaleCodeDataManager.cs
@@ -152,7 +152,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
 
         #region State Backup Methods
 
-        public string BackupAllDataBySellingNumberingPlanId(int stateBackupId, string backupDatabase, int sellingNumberPlanId)
+        public string BackupAllDataBySellingNumberingPlanId(long stateBackupId, string backupDatabase, int sellingNumberPlanId)
         {
             return String.Format(@"INSERT INTO [{0}].[TOneWhS_BE_Bkup].[SaleCode] WITH (TABLOCK)
                                             SELECT sc.[ID], sc.[Code], sc.[ZoneID], sc.[CodeGroupID], sc.[BED], sc.[EED], sc.[SourceID], {1} AS StateBackupID  FROM [TOneWhS_BE].[SaleCode]
