+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SaleZoneDataManager.cs
@@ -88,7 +88,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
 
         #region State Backup Methods
 
-        public string BackupAllDataBySellingNumberingPlanId(int stateBackupId, string backupDatabase, int sellingNumberPlanId)
+        public string BackupAllDataBySellingNumberingPlanId(long stateBackupId, string backupDatabase, int sellingNumberPlanId)
         {
             return String.Format(@"INSERT INTO [{0}].[TOneWhS_BE_Bkup].[SaleZone] WITH (TABLOCK)
                                             SELECT [ID], [SellingNumberPlanID], [CountryID], [Name], [BED], [EED], [SourceID], {1} AS StateBackupID  FROM [TOneWhS_BE].[SaleZone]
