+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SupplierZoneDataManager.cs
@@ -65,7 +65,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
 
         #region State Backup Methods
 
-        public string BackupAllDataBySupplierId(int stateBackupId, string backupDatabase, int supplierId)
+        public string BackupAllDataBySupplierId(long stateBackupId, string backupDatabase, int supplierId)
         {
             return String.Format(@"INSERT INTO [{0}].[TOneWhS_BE_Bkup].[SupplierZone] WITH (TABLOCK)
                                             SELECT [ID] ,[CountryID], [Name], [SupplierID], [BED], [EED], [SourceID],  {1} AS StateBackupID FROM [TOneWhS_BE].[SupplierZone]
