+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SupplierZoneServiceDataManager.cs
@@ -94,7 +94,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
 
         #region State Backup Methods
 
-        public string BackupAllDataBySupplierId(int stateBackupId, string backupDatabase, int supplierId)
+        public string BackupAllDataBySupplierId(long stateBackupId, string backupDatabase, int supplierId)
         {
             return String.Format(@"INSERT INTO [{0}].[TOneWhS_BE_Bkup].[SupplierZoneService] WITH (TABLOCK)
                                             SELECT zs.[ID], zs.[ZoneID], zs.[PriceListID], zs.[ReceivedServicesFlag], zs.[EffectiveServiceFlag], zs.[BED], zs.[EED], zs.[SourceID],  {1} AS StateBackupID  FROM [TOneWhS_BE].[SupplierZoneService] zs
